+++ packages/eslint-config-fisker/configs/airbnb/rules/errors.js
@@ -8,5 +8,9 @@ module.exports = {
   rules: {
     'no-empty': ['error', {allowEmptyCatch: true}],
     'no-cond-assign': ['error', 'except-parens'],
+
+    // no loop check
+    // https://eslint.org/docs/rules/no-constant-condition
+    'no-constant-condition': ['warn', {checkLoops: false}],
   },
 }