+++ packages/eslint-config-fisker/index.js
@@ -1,10 +1,14 @@
 module.exports = {
+  plugins: ['promise'],
   extends: [
     // airbnb
     './configs/airbnb/airbnb.js',
 
     // prettier
     './configs/prettier/prettier.js',
+
+    // promise
+    './configs/promise/promise.js',
 
     // own rules
     './configs/own/own.js',