+++ packages/eslint-config-fisker/lint-staged.config.js
@@ -1,3 +1,4 @@
 module.exports = {
-  '*.{js,json,md}': ['prettier --write', 'git add'],
+  '*.js': ['eslint --fix', 'prettier --write', 'git add'],
+  '*.{json,md}': ['prettier --write', 'git add'],
 }