+++ packages/eslint-config-fisker/CHANGELOG.md
@@ -1,5 +1,15 @@
 # changelog
 
+## v3.4.0
+
+- add `promise`
+- turn on `max-params` but, `warn`
+- switch `no-implicit-coercion` to `warn`
+- switch `no-constant-condition` to `warn`
+- switch `no-underscore-dangle` to `warn`
+- switch `global-require` to `off`
+- turn off `prettier/prettier`, was `warn`
+
 ## v1.0.0 ~ v3.3.1
 
 - allow no-new d9e1957